youtube/pyproject.toml
[project]
name = "arcade-youtube"
version = "0.1.0"
description = "YouTube integration toolkit for Arcade"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "google-api-python-client>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.1.0",
]
requires-python = ">=3.13"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["arcade_youtube"]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
skip-magic-trailing-comma = false
line-length = 100

[tool.poetry]
name = "arcade_youtube"
version = "0.0.1"
description = "Track watch history, subscriptions, and search activity"
authors = ["regiscamimura <regiscamimura@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.13"
arcade-sdk = "*"
google-api-python-client = "^2.0.0"
google-auth-oauthlib = "^1.0.0"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-cov = "^4.0.0"
mypy = "^1.5.1"
pre-commit = "^3.4.0"
tox = "^4.11.1"
ruff = "^0.7.4"
vcrpy = "^4.2.1"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0,<2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
files = ["arcade_youtube/**/*.py"]
python_version = "3.10"
disallow_untyped_defs = "True"
disallow_any_unimported = "True"
no_implicit_optional = "True"
check_untyped_defs = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
show_error_codes = "True"
ignore_missing_imports = "True"

[tool.pytest.ini_options]
testpaths = ["**/tests"]

[tool.coverage.report]
skip_empty = true
